tests/wpt/tests/css/css-overflow/scroll-marker-focus-visible.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Overflow Test: ::scroll-marker supports :focus-visible</title>
<link rel="help" href="https://drafts.csswg.org/css-overflow-5/#scroll-marker-pseudo">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<style>
  body {
    margin: 0;
  }

  #scroller {
    width: 600px;
    height: 300px;
    overflow: auto;
    scroll-marker-group: before;
    white-space: nowrap;
  }

  #scroller div {
    background: blue;
    display: inline-block;
    width: 600px;
    height: 270px;
  }

  #scroller::scroll-marker-group {
    height: 20px;
    width: 40px;
  }

  #scroller div::scroll-marker {
    content: "";
    width: 100px;
    height: 20px;
    background-color: red;
    display: inline-block;
  }

  #scroller div::scroll-marker:focus-visible {
    background-color: green;
  }
</style>
<div id="scroller">
  <div id="target"></div>
  <div></div>
</div>
<script>
  promise_test(async t => {
    const kTab = '\uE004';
    await new test_driver.Actions()
      .keyDown(kTab)
      .keyUp(kTab)
      .send();
    assert_equals(getComputedStyle(target, "::scroll-marker").backgroundColor, "rgb(0, 128, 0)", "::scroll-marker supports :focus-visible");
  });
</script>
