third_party/blink/web_tests/platform/mac/external/wpt/css/css-overflow/scroll-marker-focus-visible-expected.txt

This is a testharness.js-based test.
[FAIL] CSS Overflow Test: ::scroll-marker supports :focus-visible
  assert_equals: ::scroll-marker supports :focus-visible expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
Harness: the test ran to completion.
